refactor: share security severity policy via fallow-security
derive_security_severity and the catalogue-title wrapper were verbatim
copies in fallow-core and fallow-engine, so a tiering change in one
would silently diverge from the other. The policy now lives once in
fallow-security, which both crates already depend on; core's rank pass
calls it and engine re-exports it unchanged for cli/api/lsp consumers.
The severity tests move with the implementation.
